Extract reusable conversation picker from share intent screen
