+++ src/com/android/messaging/ui/conversationpicker/host/forward/ForwardMessageHandler.kt
@@ -1,22 +1,22 @@
-package com.android.messaging.ui.forward.screen
+package com.android.messaging.ui.conversationpicker.host.forward
 
 import android.app.Activity
 import com.android.messaging.data.conversation.model.draft.ConversationDraft
 import com.android.messaging.datamodel.data.MessageData
-import com.android.messaging.domain.shareintent.model.ShareSendTarget
-import com.android.messaging.domain.shareintent.usecase.SendSharedContentToTargets
+import com.android.messaging.domain.conversationpicker.model.SendTarget
+import com.android.messaging.domain.conversationpicker.usecase.SendContentToTargets
 import com.android.messaging.ui.UIIntents
-import com.android.messaging.ui.shareintent.screen.ShareIntentEffectHandler
-import com.android.messaging.ui.shareintent.screen.model.ShareIntentScreenEffect as Effect
+import com.android.messaging.ui.conversationpicker.ConversationPickerEffectHandler
+import com.android.messaging.ui.conversationpicker.model.ConversationPickerEffect as Effect
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-internal class ForwardEffectHandler(
+internal class ForwardMessageHandler(
     private val applicationScope: CoroutineScope,
     private val activity: Activity,
     private val message: MessageData,
-    private val sendSharedContentToTargets: SendSharedContentToTargets,
-) : ShareIntentEffectHandler {
+    private val sendContentToTargets: SendContentToTargets,
+) : ConversationPickerEffectHandler {
 
     override fun handle(effect: Effect) {
         when (effect) {
@@ -27,24 +27,23 @@ internal class ForwardEffectHandler(
             is Effect.SendToSelected -> {
                 sendToSelected(effect.targets, effect.draft)
             }
+
+            is Effect.OpenAttachmentPreview -> Unit
+            is Effect.OpenConversationFailed -> Unit
         }
     }
 
     private fun openConversation(conversationId: String) {
-        UIIntents.get().launchConversationActivity(
-            activity,
-            conversationId,
-            message,
-        )
+        UIIntents.get().launchConversationActivity(activity, conversationId, message)
         activity.finish()
     }
 
     private fun sendToSelected(
-        targets: Set<ShareSendTarget>,
+        targets: Set<SendTarget>,
         draft: ConversationDraft,
     ) {
         applicationScope.launch {
-            sendSharedContentToTargets(draft, targets)
+            sendContentToTargets(draft, targets)
         }
 
         UIIntents.get().launchConversationListActivity(activity)
